Revert "mostly const consistency and boolean operators"
This reverts commit 50b90a355e1c97884d0e54f6acb6ec55f979c3b4.

+++ include/Matrix.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
-#include <cstddef>
-#include <vector>
+#include <array>
 
 template <typename T>
 struct Matrix {
@@ -9,14 +8,14 @@ struct Matrix {
     int sizeX;
     int sizeY;
 
-    Matrix(int x_, int y_) : v(static_cast<std::size_t>(x_ * y_), T{}), sizeX(x_), sizeY(y_) {}
+    Matrix(int x_, int y_) : v(x_ * y_, T{}), sizeX(x_), sizeY(y_) {}
 
     T& operator()(int x_, int y_)  {
-      return v[static_cast<std::size_t>(x_ + y_ * sizeX)];
+        return v[x_ + y_ * sizeX];
     }
 
     const T& operator()(int x_, int y_) const {
-      return v[static_cast<std::size_t>(x_ + y_ * sizeX)];
+        return v[x_ + y_ * sizeX];
     }
 };
